+++ src/femlib/GenericMesh.hpp
@@ -581,7 +581,7 @@ public:
     return  BuildDFNumbering(dfon,nbequibe,equibe);
   }
   int nElementonB(int k,int j) const // add v4 F.H 12/2018 (not sure for internal boundary !!!)
-    { int kk= TheAdjacencesLink[nea*k+j]; return (kk>=0) && (kk%nea  != k) ? 2 : 1;}
+    { int kk= TheAdjacencesLink[nea*k+j]; return (kk>=0) && (kk/nea  != k) ? 2 : 1;}// correct FH 23 mrch 2012
 
   int ElementAdj(int k,int &j) const  {
     int p=TheAdjacencesLink[nea*k+j];